Enable database for multisite
Basically to install the plugin to a vanilla multi-site this should fix everything. For upgrades or new multi-site it looks like it should also be good as each site gets it's own options table. Lastly in the case that there are multiple sites, on multiple versions of the plugin through some form of arcane black-magic, this should be able to correct the schema, just as well as the plugin did for a single site before changes were made.

 * update made a thin wrapper for logic
 * old update code moved to new private `site_do_update` method

All old tests should work still